requirements.txt
streamlit
pandas
pymupdf
pandas
streamlit_aggrid     
Pillow
fuzzywuzzy
openpyxl
azure-storage-file-datalake
pdf2image
azure-identity
azure-storage-blob
fsspec
adlfs
pyodbc
pyinstrument
altair<5
sqlalchemy
streamlit_antd_components